+++ randombytes.h
@@ -1,1 +1,1 @@
-extern void randombytes(unsigned char * ptr,unsigned int length);
+extern void randombytes(unsigned char *ptr,unsigned long long length);
